Remove content_typed namespace.
- content_typed.py -> models.py + admin.py
- content_typed/change_list.html ->
  admin/includes/content_typed_change_list.html
