+++ mezzanine/core/templates/admin/includes/content_typed_change_list.html
